first pass of moving logic into extendedbundle
