--- pkg/cnab/dependencies/v1/doc.go
@@ -1,2 +1,0 @@
-// Package v1 contains the implementation for v1 of the CNAB Dependencies specification.
-package v1
