--- pkg/cnab/dependencies/v1/installations.go
@@ -1,10 +1,0 @@
-package v1
-
-import (
-	"fmt"
-)
-
-// BuildPrerequisiteInstallationName generates the name of a prerequisite dependency installation.
-func BuildPrerequisiteInstallationName(installation string, dependency string) string {
-	return fmt.Sprintf("%s-%s", installation, dependency)
-}
