+++ pkg/cnab/extended_bundle_test.go
@@ -3,6 +3,7 @@ package cnab
 import (
 	"testing"
 
+	depsv1ext "get.porter.sh/porter/pkg/cnab/extensions/dependencies/v1"
 	"get.porter.sh/porter/pkg/portercontext"
 	porterschema "get.porter.sh/porter/pkg/schema"
 	"github.com/cnabio/cnab-go/bundle"
@@ -238,6 +239,95 @@ func TestValidate(t *testing.T) {
 			}
 			require.NoError(t, err)
 			require.Contains(t, cxt.GetError(), tc.wantErr)
+		})
+	}
+}
+
+func TestExtendedBundle_ResolveDependencies(t *testing.T) {
+	t.Parallel()
+
+	bun := NewBundle(bundle.Bundle{
+		Custom: map[string]interface{}{
+			DependenciesV1ExtensionKey: depsv1ext.Dependencies{
+				Requires: map[string]depsv1ext.Dependency{
+					"mysql": {
+						Bundle: "getporter/mysql:5.7",
+					},
+					"nginx": {
+						Bundle: "localhost:5000/nginx:1.19",
+					},
+				},
+			},
+		},
+	})
+
+	eb := ExtendedBundle{}
+	locks, err := eb.ResolveDependencies(bun)
+	require.NoError(t, err)
+	require.Len(t, locks, 2)
+
+	var mysql DependencyLock
+	var nginx DependencyLock
+	for _, lock := range locks {
+		switch lock.Alias {
+		case "mysql":
+			mysql = lock
+		case "nginx":
+			nginx = lock
+		}
+	}
+
+	assert.Equal(t, "getporter/mysql:5.7", mysql.Reference)
+	assert.Equal(t, "localhost:5000/nginx:1.19", nginx.Reference)
+}
+
+func TestExtendedBundle_ResolveVersion(t *testing.T) {
+	t.Parallel()
+
+	testcases := []struct {
+		name        string
+		dep         depsv1ext.Dependency
+		wantVersion string
+		wantError   string
+	}{
+		{name: "pinned version",
+			dep:         depsv1ext.Dependency{Bundle: "mysql:5.7"},
+			wantVersion: "5.7"},
+		{name: "unimplemented range",
+			dep:       depsv1ext.Dependency{Bundle: "mysql", Version: &depsv1ext.DependencyVersion{Ranges: []string{"1 - 1.5"}}},
+			wantError: "not implemented"},
+		{name: "default tag to latest",
+			dep:         depsv1ext.Dependency{Bundle: "getporterci/porter-test-only-latest"},
+			wantVersion: "latest"},
+		{name: "no default tag",
+			dep:       depsv1ext.Dependency{Bundle: "getporterci/porter-test-no-default-tag"},
+			wantError: "no tag was specified"},
+		{name: "default tag to highest semver",
+			dep:         depsv1ext.Dependency{Bundle: "getporterci/porter-test-with-versions", Version: &depsv1ext.DependencyVersion{Ranges: nil, AllowPrereleases: true}},
+			wantVersion: "v1.3-beta1"},
+		{name: "default tag to highest semver, explicitly excluding prereleases",
+			dep:         depsv1ext.Dependency{Bundle: "getporterci/porter-test-with-versions", Version: &depsv1ext.DependencyVersion{Ranges: nil, AllowPrereleases: false}},
+			wantVersion: "v1.2"},
+		{name: "default tag to highest semver, excluding prereleases by default",
+			dep:         depsv1ext.Dependency{Bundle: "getporterci/porter-test-with-versions"},
+			wantVersion: "v1.2"},
+	}
+
+	for _, tc := range testcases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
+			eb := ExtendedBundle{}
+			version, err := eb.ResolveVersion("mysql", tc.dep)
+			if tc.wantError != "" {
+				require.Error(t, err, "ResolveVersion should have returned an error")
+				assert.Contains(t, err.Error(), tc.wantError)
+			} else {
+				require.NoError(t, err, "ResolveVersion should not have returned an error")
+
+				assert.Equal(t, tc.wantVersion, version.Tag(), "incorrect version resolved")
+			}
 		})
 	}
 }
